+++ src/Pages/Dashboard/Components/BlockSpotlight.js
@@ -7,7 +7,6 @@ import { BlockImage, Content, Loading, Sprite as BaseSprite, DataCard } from 'Co
 import { maxLength, getUTCTime, numberFormat, formatSeconds, formatDenom } from 'utils';
 import { polling } from 'consts';
 import useOrderbook from 'redux/hooks/useOrderbook';
-import isYesterday from 'date-fns/isYesterday';
 
 const Group = styled.div`
   display: flex;
@@ -35,26 +34,7 @@ const BlockSpotlight = () => {
   const [blockLoading, setBlockLoading] = useState(false);
   const { blockLatest, getBlockSpotlight, blockSpotlightFailed, blockSpotlightLoading } =
     useBlocks();
-  const { dailyPrice, priceHistory, getDailyPrice, getPriceHistory } = useOrderbook();
-
-  // Find all the prev day's prices and get the average price for the previous day
-  const prevDayPrices = priceHistory.filter(i => isYesterday(new Date(i.dateTime)));
-  const prevDayAverage = prevDayPrices
-    .reduce((acc, curr) => acc.add(curr.displayPricePerDisplayUnit), new Big(0))
-    .div(prevDayPrices.length || 1);
-
-  const latestDisplayPrice = dailyPrice.latestDisplayPricePerDisplayUnit || 1;
-  // Check if the price increased from yesterday
-  const priceIncrease = prevDayAverage.lte(latestDisplayPrice);
-  // figure out the price change from yesterday
-  const priceChange = priceIncrease
-    ? prevDayAverage.div(latestDisplayPrice)
-    : new Big(latestDisplayPrice).div(prevDayAverage);
-  // create percentage based on the change
-  const priceChangePercent = `${priceIncrease ? `+` : `-`}${new Big(1)
-    .minus(priceChange)
-    .times(100)
-    .toFixed(0)}%`;
+  const { dailyPrice, getDailyPrice } = useOrderbook();
 
   // Initial load, get most recent blocks
   useEffect(() => {
@@ -64,14 +44,14 @@ const BlockSpotlight = () => {
         setInitialLoad(false);
         // Get initial blocks
         try {
-          await Promise.all([getBlockSpotlight(), getDailyPrice(), getPriceHistory('WEEK')]);
+          await Promise.all([getBlockSpotlight(), getDailyPrice()]);
           setBlockLoading(false);
         } catch (e) {
           setBlockLoading(false);
         }
       }
     })();
-  }, [getBlockSpotlight, getDailyPrice, getPriceHistory, initialLoad]);
+  }, [getBlockSpotlight, getDailyPrice, initialLoad]);
 
   // Poll the API for new data every 5s
   useInterval(
@@ -110,7 +90,6 @@ const BlockSpotlight = () => {
 
   const latestPrice = new Big(dailyPrice.latestDisplayPricePerDisplayUnit || 0);
   const twentyFourHourVolume = latestPrice.times(dailyPrice.displayVolumeTraded || 0);
-  const marketCap = latestPrice.times(100000000000);
 
   return (
     <Content justify="center" alignItems="flex-start">
@@ -162,17 +141,6 @@ const BlockSpotlight = () => {
             </DataCard>
             <DataCard icon="CALENDAR" title="24hr Volume">
               {`$${formatDenom(twentyFourHourVolume, 'USD', {
-                shorthand: true,
-                decimal: 2,
-              })}`}
-            </DataCard>
-            <DataCard icon="PRICE" title="Latest Price">
-              {`$${formatDenom(dailyPrice.latestDisplayPricePerDisplayUnit, 'USD', {
-                minimumFractionDigits: 3,
-              })} (${priceChangePercent})`}
-            </DataCard>
-            <DataCard icon="LINE_CHART" title="Market Cap">
-              {`$${formatDenom(marketCap, 'USD', {
                 shorthand: true,
                 decimal: 2,
               })}`}